Обновление проектов. ничего особо не изменилось. перенастраивал проекты просто
