+++ RESTfulAPIKorpPortal/src/main/java/by/bsuir/ief/rest/model/entity/Calendar.java
@@ -1,9 +1,9 @@
 package by.bsuir.ief.rest.model.entity;
 
-import org.hibernate.annotations.Type;
-//import org.joda.time.DateTime;
-
 import javax.persistence.*;
+import java.util.Date;
+
+//import org.joda.time.DateTime;
 
 /**
  * Created by andrey on 04.04.2016.
@@ -14,7 +14,7 @@ public class Calendar implements Cloneable {
     private Integer idCalendar;
     private Integer week;
     private Integer quartal;
-    private java.util.Date date;
+    private Date date;
 
     @Id
     @Column(name = "id_calendar", nullable = false)
@@ -50,11 +50,11 @@ public class Calendar implements Cloneable {
     @Basic
     @Column(name = "date", nullable = false)
     @Temporal(TemporalType.DATE)
-    public java.util.Date getDate() {
+    public Date getDate() {
         return date;
     }
 
-    public void setDate(java.util.Date date) {
+    public void setDate(Date date) {
         this.date = date;
     }
