+++ RESTfulAPIKorpPortal/src/main/resources/log4j.properties
@@ -6,7 +6,7 @@ log4j.rootLogger=INFO, file, stdout
 
 # Direct log messages to a log file
 log4j.appender.file=org.apache.log4j.RollingFileAppender
-log4j.appender.file.File=D:\\app_log.log
+log4j.appender.file.File=D:\\app_server.log
 log4j.appender.file.MaxFileSize=100MB
 log4j.appender.file.MaxBackupIndex=7
 log4j.appender.file.layout=org.apache.log4j.PatternLayout
