+++ WebCorporativPortal/pom.xml
@@ -17,6 +17,7 @@
         <hibernate-version>5.1.0.Final</hibernate-version>
         <jackson.databind-version>2.7.3</jackson.databind-version>
         <joda-time>2.9.3</joda-time>
+        <cglib>3.2.2</cglib>
         <org.jadira.usertype>5.0.0.GA</org.jadira.usertype>
     </properties>
 
@@ -31,6 +32,31 @@
             <groupId>org.springframework</groupId>
             <artifactId>spring-webmvc</artifactId>
             <version>${org.springframework-version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-aspects</artifactId>
+            <version>${org.springframework-version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-test</artifactId>
+            <version>${org.springframework-version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-web</artifactId>
+            <version>${org.springframework-version}</version>
+        </dependency>
+
+
+        <dependency>
+            <groupId>cglib</groupId>
+            <artifactId>cglib</artifactId>
+            <version>${cglib}</version>
         </dependency>
 
         <!-- AspectJ -->
@@ -142,11 +168,6 @@
             <artifactId>junit</artifactId>
             <version>4.7</version>
             <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
-            <artifactId>jackson-databind</artifactId>
-            <version>2.7.3</version>
         </dependency>
 
         <!-- joda-time -->
