--- WebCorporativPortal/src/main/java/by/bsuir/ief/corporativ_portal/controller/LoginController.java
@@ -1,55 +1,0 @@
-package by.bsuir.ief.corporativ_portal.controller;
-
-import by.bsuir.ief.corporativ_portal.model.configue.ClientURL;
-import by.bsuir.ief.corporativ_portal.model.entity.User;
-import by.bsuir.ief.corporativ_portal.model.service.ServiceManager;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.servlet.ModelAndView;
-
-import javax.servlet.http.HttpSession;
-import javax.validation.Valid;
-
-@Controller
-@RequestMapping("/")
-public class LoginController {
-
-    private static final Logger logger = LoggerFactory.getLogger(LoginController.class);
-
-    @Qualifier("serviceManager")
-    @Autowired
-    private ServiceManager serviceManager;
-
-    @RequestMapping(value="/", method = RequestMethod.GET)
-    public ModelAndView main(HttpSession session){
-        session.invalidate();
-        return new ModelAndView(ClientURL.getProperty("url.index")/*"index"*/, "user", new User());
-    }
-
-    @RequestMapping(value = "/check-user", method = RequestMethod.POST)
-    public String checkUser(@Valid @ModelAttribute("user") User user, BindingResult bindingResult, Model model, HttpSession session) {
-        if (bindingResult.hasErrors()) {
-            return ClientURL.getProperty("url.login");
-        }
-        try {
-            user = serviceManager.getUserByLogin(user);
-            model.addAttribute("user", user);
-            model.addAttribute("person", user.getPerson());
-            session.setAttribute("user", user);
-            return ClientURL.getProperty("url.main");
-        } catch (Exception e) {
-            e.printStackTrace();
-            return ClientURL.getProperty("url.error.wrongloginorpassword");
-        }
-    }
-
-
-}
