--- WebCorporativPortal/src/main/java/by/bsuir/ief/corporativ_portal/controller/MainPageController.java
@@ -1,7 +1,0 @@
-package by.bsuir.ief.corporativ_portal.controller;
-
-/**
- * Created by Darya on 29.04.16.
- */
-public class MainPageController {
-}
