+++ WebCorporativPortal/src/main/java/by/bsuir/ief/corporativ_portal/model/server/ConnectionServer.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
 
+
 @Component
 public class ConnectionServer {
