+++ WebCorporativPortal/src/main/java/by/bsuir/ief/corporativ_portal/model/service/UserService.java
@@ -1,0 +1,13 @@
+package by.bsuir.ief.corporativ_portal.model.service;
+
+import org.springframework.stereotype.Component;
+
+/**
+ * Created by andrey on 04.05.2016.
+ */
+@Component
+public class UserService {
+
+
+
+}
