--- client7REST/.idea/artifacts/client7REST_war_exploded.xml
@@ -1,58 +1,0 @@
-<component name="ArtifactManager">
-  <artifact type="exploded-war" name="client7REST:war exploded">
-    <output-path>$PROJECT_DIR$/target/lb7-1.0.0-BUILD-SNAPSHOT</output-path>
-    <root id="root">
-      <element id="directory" name="WEB-INF">
-        <element id="directory" name="classes">
-          <element id="module-output" name="client7REST" />
-        </element>
-        <element id="directory" name="lib">
-          <element id="library" level="project" name="Maven: org.springframework:spring-context:4.2.5.RELEASE" />
-          <element id="library" level="project" name="Maven: org.springframework:spring-aop:4.2.5.RELEASE" />
-          <element id="library" level="project" name="Maven: aopalliance:aopalliance:1.0" />
-          <element id="library" level="project" name="Maven: org.springframework:spring-beans:4.2.5.RELEASE" />
-          <element id="library" level="project" name="Maven: org.springframework:spring-core:4.2.5.RELEASE" />
-          <element id="library" level="project" name="Maven: commons-logging:commons-logging:1.2" />
-          <element id="library" level="project" name="Maven: org.springframework:spring-expression:4.2.5.RELEASE" />
-          <element id="library" level="project" name="Maven: org.springframework:spring-webmvc:4.2.5.RELEASE" />
-          <element id="library" level="project" name="Maven: org.springframework:spring-web:4.2.5.RELEASE" />
-          <element id="library" level="project" name="Maven: org.aspectj:aspectjrt:1.8.0" />
-          <element id="library" level="project" name="Maven: javax.inject:javax.inject:1" />
-          <element id="library" level="project" name="Maven: javax.servlet:jstl:1.2" />
-          <element id="library" level="project" name="Maven: org.slf4j:slf4j-api:1.7.7" />
-          <element id="library" level="project" name="Maven: org.slf4j:jcl-over-slf4j:1.7.7" />
-          <element id="library" level="project" name="Maven: org.slf4j:slf4j-log4j12:1.7.7" />
-          <element id="library" level="project" name="Maven: log4j:log4j:1.2.17" />
-          <element id="library" level="project" name="Maven: org.codehaus.jackson:jackson-mapper-asl:1.9.13" />
-          <element id="library" level="project" name="Maven: org.codehaus.jackson:jackson-core-asl:1.9.13" />
-          <element id="library" level="project" name="Maven: org.webjars:bootstrap:3.2.0" />
-          <element id="library" level="project" name="Maven: org.webjars:jquery:1.11.1" />
-          <element id="library" level="project" name="Maven: org.hibernate:hibernate-validator:5.2.4.Final" />
-          <element id="library" level="project" name="Maven: javax.validation:validation-api:1.1.0.Final" />
-          <element id="library" level="project" name="Maven: org.jboss.logging:jboss-logging:3.2.1.Final" />
-          <element id="library" level="project" name="Maven: com.fasterxml:classmate:1.1.0" />
-          <element id="library" level="project" name="Maven: com.fasterxml.jackson.core:jackson-databind:2.7.3" />
-          <element id="library" level="project" name="Maven: com.fasterxml.jackson.core:jackson-annotations:2.7.0" />
-          <element id="library" level="project" name="Maven: com.fasterxml.jackson.core:jackson-core:2.7.3" />
-          <element id="library" level="project" name="Maven: joda-time:joda-time:2.9.3" />
-          <element id="library" level="project" name="Maven: org.jadira.usertype:usertype.core:5.0.0.GA" />
-          <element id="library" level="project" name="Maven: org.hibernate:hibernate-entitymanager:5.0.1.Final" />
-          <element id="library" level="project" name="Maven: org.hibernate:hibernate-core:5.0.1.Final" />
-          <element id="library" level="project" name="Maven: antlr:antlr:2.7.7" />
-          <element id="library" level="project" name="Maven: org.jboss:jandex:1.2.2.Final" />
-          <element id="library" level="project" name="Maven: dom4j:dom4j:1.6.1" />
-          <element id="library" level="project" name="Maven: xml-apis:xml-apis:1.0.b2" />
-          <element id="library" level="project" name="Maven: org.hibernate.common:hibernate-commons-annotations:5.0.0.Final" />
-          <element id="library" level="project" name="Maven: org.hibernate.javax.persistence:hibernate-jpa-2.1-api:1.0.0.Final" />
-          <element id="library" level="project" name="Maven: org.apache.geronimo.specs:geronimo-jta_1.1_spec:1.1.1" />
-          <element id="library" level="project" name="Maven: org.javassist:javassist:3.18.1-GA" />
-          <element id="library" level="project" name="Maven: org.jadira.usertype:usertype.spi:5.0.0.GA" />
-        </element>
-      </element>
-      <element id="directory" name="META-INF">
-        <element id="file-copy" path="$PROJECT_DIR$/target/lb7-1.0.0-BUILD-SNAPSHOT/META-INF/MANIFEST.MF" />
-      </element>
-      <element id="javaee-facet-resources" facet="client7REST/web/Web" />
-    </root>
-  </artifact>
-</component>
