--- client7REST/src/main/webapp/WEB-INF/views/index.jsp
@@ -1,9 +1,0 @@
-<%@ page contentType="text/html;charset=UTF-8" language="java" %>
-<html>
-<head>
-    <title>Index</title>
-</head>
-<body>
-<jsp:forward page="table.jsp"/>
-</body>
-</html>
