--- clientPisl/src/main/java/by/bsuir/pisl/controller/Controller.java
@@ -1,8 +1,0 @@
-package by.bsuir.pisl.controller;
-
-/**
- * Created by andrey on 20.04.2016.
- */
-public interface Controller {
-
-}
