--- clientPisl/src/main/java/by/bsuir/pisl/controller/DialogManager.java
@@ -1,33 +1,0 @@
-package by.bsuir.pisl.controller;
-
-import javafx.scene.control.Alert;
-
-/**
- * Created by lumiandrey on 07.12.2015.
- */
-public class DialogManager {
-    public static void showInfoDialog(String title, String text){
-        Alert alert = new Alert(Alert.AlertType.INFORMATION);
-        alert.setTitle(title);
-        alert.setContentText(text);
-        alert.setHeaderText("");
-        alert.showAndWait();
-    }
-
-    public static void showErrorDialog(String title, String text){
-        Alert alert = new Alert(Alert.AlertType.ERROR);
-        alert.setTitle(title);
-        alert.setContentText(text);
-        alert.setHeaderText("");
-        alert.showAndWait();
-    }
-    public static void showWarningDialog(String title, String text)
-    {
-        Alert alert = new Alert(Alert.AlertType.WARNING);
-        alert.setTitle(title);
-        alert.setContentText(text);
-        alert.setHeaderText("");
-        alert.showAndWait();
-
-    }
-}
