--- clientPisl/src/main/java/by/bsuir/pisl/example
@@ -1,26 +1,0 @@
-package ru.habrahabr;
-import javafx.application.Application;
-import org.springframework.boot.SpringApplication;
-import org.springframework.context.ConfigurableApplicationContext;
-
-public abstract class AbstractJavaFxApplicationSupport extends Application
-{
- private static String[] savedArgs;
-  protected ConfigurableApplicationContext context;
-     @Override
-     public void init() throws Exception {
-          context = SpringApplication.run(getClass(), savedArgs);
-                 context.getAutowireCapableBeanFactory().autowireBean(this);
-     }
-
-      @Override
-      public void stop() throws Exception {
-            super.stop();         context.close();
-      }
-
-      protected static void launchApp(Class<? extends AbstractJavaFxApplicationSupport> clazz, String[] args)
-      {
-         AbstractJavaFxApplicationSupport.savedArgs = args;
-         Application.launch(clazz, args);
-      }
-}
