--- clientPisl/src/main/java/by/bsuir/pisl/model/ConnectionServer.java
@@ -1,110 +1,0 @@
-package by.bsuir.pisl.model;
-
-import by.bsuir.pisl.model.entity.PersonPisl;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Component;
-import org.springframework.web.client.RestTemplate;
-
-import java.net.URI;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-
-/**
- * Created by andrey on 19.04.2016.
- */
-@Component
-public class ConnectionServer {
-
-    @Autowired
-    private RestTemplate restTemplate;
-
-    private static String NAME_HOST = "http://localhost:8080/restCorpPortal/v1";
-
-    private static String GET_PERSONS = "/pislrest/persons";
-    private static String GET_PERSON_BY_ID = "/pislrest/person/{id}";
-
-    private static String POST_SAVE_PERSON = "/pislrest/person";
-    private static String POST_SAVE_PERSONS = "/pislrest/persons";
-
-    private static String POST_SAVE_OR_UPDATE_PERSON = "/pislrest/personsaveorupdate";
-    private static String POST_SAVES_OR_UPDATES_PERSONS = "/pislrest/personsaveorupdates";
-
-    private static String DELETE_PERSON_BY_ID = "/pislrest/person/{id}";
-    private static String DELETE_ALL_PERSON = "/pislrest/persons";
-
-    private static String PUT_PERSON = "/pislrest/person";
-
-    private static String GET_STATUS_SERVER = "/option/";
-
-
-
-    public List<PersonPisl> getAllPerson()throws Exception
-    {
-        //PersonPisl[] personPisls = restTemplate.getForObject(NAME_HOST+GET_PERSONS,PersonPisl[].class);
-        ResponseEntity<PersonPisl[]> responseEntity = restTemplate.getForEntity(new URI(NAME_HOST+GET_PERSONS),PersonPisl[].class);
-        System.out.println(responseEntity.getStatusCode());
-        List<PersonPisl> pislList = new ArrayList<>();
-        Collections.addAll(pislList, responseEntity.getBody());
-        return pislList;
-    }
-
-    public PersonPisl getPersonById(int id) throws Exception
-    {
-        ResponseEntity<PersonPisl> responseEntity = restTemplate.getForEntity(NAME_HOST+GET_PERSON_BY_ID,PersonPisl.class,id);
-
-        System.out.println(responseEntity.getStatusCode());
-        return responseEntity.getBody();
-    }
-
-    public PersonPisl saveOrUpdate(PersonPisl personPisl)throws Exception {
-        personPisl = restTemplate.postForObject(NAME_HOST + POST_SAVE_OR_UPDATE_PERSON, personPisl, PersonPisl.class);
-        return personPisl;
-    }
-
-    public void deletePersonById(int id)throws Exception //тут тоже
-    {
-        restTemplate.delete(NAME_HOST+DELETE_PERSON_BY_ID,new Integer(id));
-    }
-
-    public void allDelete() throws Exception
-    {
-        restTemplate.delete(new URI(NAME_HOST+DELETE_ALL_PERSON));
-    }
-
-    public List<PersonPisl> saveOrUpdate(List<PersonPisl> personPisls)throws Exception
-    {
-        personPisls = restTemplate.postForObject(NAME_HOST+POST_SAVES_OR_UPDATES_PERSONS,personPisls,List.class);
-        return personPisls;
-    }
-
-    public void putPerson(PersonPisl personPisl){
-        restTemplate.put(NAME_HOST+PUT_PERSON,personPisl);
-    }
-
-    public void putPerson(List<PersonPisl> personPisls){
-        restTemplate.put(NAME_HOST+PUT_PERSON,personPisls);
-    }
-
-    public void getStatus(){
-        int status = restTemplate.getForObject(NAME_HOST+GET_STATUS_SERVER,Integer.class);
-        if(status != 200)
-            throw new RuntimeException();
-    }
-
-
-    public PersonPisl save(PersonPisl personPisl)
-    {
-        return restTemplate.postForObject(NAME_HOST+POST_SAVE_PERSON,personPisl,PersonPisl.class);
-    }
-
-    public List<PersonPisl> save(List<PersonPisl> personPisl)
-    {
-        PersonPisl[] add = restTemplate.postForObject(NAME_HOST+POST_SAVE_PERSONS, personPisl.toArray(),PersonPisl[].class);
-        List<PersonPisl> pislList = new ArrayList<>();
-        Collections.addAll(pislList, add);
-        return pislList;
-    }
-
-}
