--- clientPisl/src/main/java/by/bsuir/pisl/model/data/PersonPislDAO.java
@@ -1,23 +1,0 @@
-package by.bsuir.pisl.model.data;
-
-
-import by.bsuir.pisl.model.entity.PersonPisl;
-
-import java.util.List;
-
-/**
- * Created by andrey on 12.04.2016.
- */
-public interface PersonPislDAO {
-    public PersonPisl addEntity(PersonPisl PersonPisl) throws Exception;
-    public List<PersonPisl> addEntitys(List<PersonPisl> personPisls) throws Exception;
-    public PersonPisl updateEntity(PersonPisl personPisl) throws Exception;
-    public List<PersonPisl> updateEntitys(List<PersonPisl> personPisls) throws Exception;
-    public PersonPisl getEntityById(int id) throws Exception;
-    public List getEntityList() throws Exception;
-    public List<PersonPisl> saveOrUpdates(List<PersonPisl> personPisls) throws Exception;
-    public PersonPisl saveOrUpdates(PersonPisl personPisl) throws Exception;
-    public boolean deleteEntity(PersonPisl personPisl) throws Exception;
-    public boolean deleteEntity(int id) throws Exception;
-    public boolean deleteAllEntity() throws Exception;
-}
