--- clientPisl/src/main/java/by/bsuir/pisl/model/entity/enumeration/Cities.java
@@ -1,21 +1,0 @@
-package by.bsuir.pisl.model.entity.enumeration;
-
-/**
- * Created by andrey on 12.04.2016.
- */
-public enum Cities {
-    Minsk,
-    Budapest,
-    Oslo,
-    Amsterdam,
-    Brest,
-    Grodno,
-    Madrid,
-    Kiev,
-    Moscow,
-    Bitebsk,
-    Warsawa,
-    Mogilev,
-    Orsha,
-    Gomel
-}
