--- clientPisl/src/main/java/by/bsuir/pisl/model/entity/enumeration/Disability.java
@@ -1,15 +1,0 @@
-package by.bsuir.pisl.model.entity.enumeration;
-
-/**
- * Created by andrey on 12.04.2016.
- */
-public enum Disability {
-    Not,
-    FirstGroup,
-    SecondGroup,
-    ThirdGroup,
-    Нет,
-    Первая_группа,
-    Вторая_группа,
-    Третья_группа
-}
