--- clientPisl/src/main/java/by/bsuir/pisl/model/entity/enumeration/MaritalStatus.java
@@ -1,13 +1,0 @@
-package by.bsuir.pisl.model.entity.enumeration;
-
-/**
- * Created by andrey on 12.04.2016.
- */
-public enum MaritalStatus {
-    Not,
-    Merried,
-    Женат,
-    Замужем,
-    Холостой,
-    не_в_отношениях
-}
