--- clientPisl/src/main/java/by/bsuir/pisl/model/entity/enumeration/Nationality.java
@@ -1,14 +1,0 @@
-package by.bsuir.pisl.model.entity.enumeration;
-
-/**
- * Created by andrey on 12.04.2016.
- */
-public enum Nationality {
-    Belarus,
-    Russia,
-    Norway,
-    Niderlands,
-    Ukrain,
-    Spain,
-    Vengria
-}
