--- clientPisl/src/main/java/by/bsuir/pisl/model/except/badexception/BadGetEntityException.java
@@ -1,15 +1,0 @@
-package by.bsuir.pisl.model.except.badexception;
-
-/**
- * Created by andrey on 17.04.2016.
- */
-public class BadGetEntityException extends BadEntityException {
-    public BadGetEntityException( String nameClassEntity, Exception ex) {
-        super("Error get Entity to base. Entity : " + nameClassEntity + " Declaration error: " + ex.getMessage());
-    }
-
-    @Override
-    public String toString() {
-        return super.toString();
-    }
-}
