--- clientPisl/src/main/resources/hibernate/hibernate-configuration.xml
@@ -1,14 +1,0 @@
-<?xml version='1.0' encoding='utf-8'?>
-<!DOCTYPE hibernate-configuration PUBLIC
-    "-//Hibernate/Hibernate Configuration DTD//EN"
-    "http://www.hibernate.org/dtd/hibernate-configuration-3.0.dtd">
-<hibernate-configuration>
-      <session-factory>
-            <!--<property name="connection.url">jdbc:mysql://localhost:3306/korporativ_portal</property>
-            <property name="connection.driver_class">com.mysql.jdbc.Driver</property>-->
-
-            <mapping class="by.bsuir.pisl.model.entity.PersonPisl"/>
-            <!-- DB schema will be updated if needed -->
-            <!-- <property name="hbm2ddl.auto">update</property> -->
-      </session-factory>
-</hibernate-configuration>
